pulling out custom hook & refactoring layout to not be so wonky
