+++ components/layout.tsx
@@ -1,13 +1,12 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import Head from "next/head";
 
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { grey, blue } from "@mui/material/colors";
 
-import { NavbarDesktop, NavbarMobileLarge, NavbarMobileSmall } from "./navbar";
-
+import { NavbarDesktop, NavbarMobileSmall } from "./navbar";
+import useIsMobile from "./useIsMobile";
 import styles from "./layout.module.scss";
-import { layoutBreakpoint } from "../styles/_globalvariables.module.scss";
 
 //***** Theme  *****/
 const theme = createTheme({
@@ -33,18 +32,11 @@ const HeadComponent = () => (
 );
 
 //***** Mobile Layout *****/
-type NavbarStyle = "mobileLarge" | "mobileSmall";
-
-function MobileLayout({ children, navbarStyle }: { children: React.ReactNode; navbarStyle?: NavbarStyle }) {
-  let navbarMap = {
-    mobileLarge: <NavbarMobileLarge />,
-    mobileSmall: <NavbarMobileSmall />,
-  };
-  let Navbar = navbarStyle ? navbarMap[navbarStyle] : "";
+function MobileLayout({ children }: { children: React.ReactNode }) {
   return (
     <ThemeProvider theme={theme}>
       <HeadComponent />
-      {Navbar}
+      <NavbarMobileSmall />
       <main className={styles.mobilePageContainer}>{children}</main>
     </ThemeProvider>
   );
@@ -61,46 +53,23 @@ function DesktopLayout({ children }: { children: React.ReactNode }) {
   );
 }
 
-//***** Window Resize Handling  *****/
-
-//debounce on resize so that performance issues aren't caused by
-//ppl going bonkers on resizing their windows
-const debounceInterval = 10; //in ms - increase me if performance issues occur
-const debounce = (fn: Function, ms = debounceInterval) => {
-  let timeoutId: ReturnType<typeof setTimeout>;
-  return function (this: any, ...args: any[]) {
-    clearTimeout(timeoutId);
-    timeoutId = setTimeout(() => fn.apply(this, args), ms);
-  };
-};
-
-//returns correct layout initially and (after a debounce) on every resize event
-const resizeBreakpoint = parseInt(layoutBreakpoint.replace("px", "")); //imported from scss
-function ResponsiveResize() {
-  const [mobile, setMobile] = useState<boolean | undefined>(undefined);
-
-  useEffect(() => {
-    let isMounted: boolean = true;
-    const updateMobile = () => {
-      if (isMounted)
-        setMobile(
-          //window.visualViewport is so resizing works correctly in Chrome Devtools
-          window.innerWidth < resizeBreakpoint || window.visualViewport.width < resizeBreakpoint ? true : false
-        );
-    };
-
-    updateMobile(); //not debounced -- initial calc
-    window.addEventListener("resize", debounce(updateMobile)); //debounced!
-    return () => {
-      window.removeEventListener("resize", debounce(updateMobile)); //debounced!
-      isMounted = false;
-    };
-  }, []);
-  return typeof mobile !== "undefined" ? (mobile ? MobileLayout : DesktopLayout) : null;
-}
-
 //returns the correct component (MobileLayout or Desktop Layout) based on current window size
-export default function Layout({ children, navbarStyle }: { children: React.ReactNode; navbarStyle?: NavbarStyle }) {
-  let ResponsiveLayout = ResponsiveResize();
-  return ResponsiveLayout ? ResponsiveLayout({ children, navbarStyle }) : null;
+export function Layout({ children }: { children: React.ReactNode }) {
+  return useIsMobile() ? MobileLayout({ children }) : DesktopLayout({ children });
 }
+
+//searchpage
+//desktop
+// (<ThemeProvider theme={theme}>
+// <HeadComponent />
+// <NavbarDesktop />
+// <main className={styles.searchPageContainer}>{children}</main>
+// </ThemeProvider>)
+//mobile
+// (
+//   <ThemeProvider theme={theme}>
+//     <HeadComponent />
+//     <NavbarMobileLarge />
+//     <main className={styles.mobilePageContainer}>{children}</main>
+//   </ThemeProvider>
+// )
