+++ pages/contact.tsx
@@ -1,10 +1,10 @@
 import type { NextPage } from "next";
 
-import Layout from "../components/layout";
+import { Layout } from "../components/layout";
 
 const Contact: NextPage = () => {
   return (
-    <Layout navbarStyle="mobileSmall">
+    <Layout>
       <h1 className="title">Say Hi!</h1>
       <p className="bodyText">
         Have a bug to report? Incredibly upset with how we parse scent descriptions? Want to shower us in praise for
