+++ pages/houses.tsx
@@ -1,10 +1,10 @@
 import type { NextPage } from "next";
 
-import Layout from "../components/layout";
+import { Layout } from "../components/layout";
 
 const AllHouses: NextPage = () => {
   return (
-    <Layout navbarStyle="mobileSmall">
+    <Layout>
       <h1 className="title">All Houses</h1>
     </Layout>
   );
