+++ pages/houses/[house].tsx
@@ -1,12 +1,12 @@
 import { useRouter } from "next/router";
 
-import Layout from "../../components/layout";
+import { Layout } from "../../components/layout";
 
 const House = () => {
   const router = useRouter();
   const { house } = router.query;
   return (
-    <Layout navbarStyle="mobileSmall">
+    <Layout>
       <h1 className="title">House: {house}</h1>
     </Layout>
   );
