+++ pages/scents/[scent].tsx
@@ -1,12 +1,12 @@
 import { useRouter } from "next/router";
 
-import Layout from "../../components/layout";
+import { Layout } from "../../components/layout";
 
 const Scent = () => {
   const router = useRouter();
   const { scent } = router.query;
   return (
-    <Layout navbarStyle="mobileSmall">
+    <Layout>
       <h1 className="title">Scent: {scent}</h1>
     </Layout>
   );
